# swagger.yaml
openapi: 3.0.0
info:
  title: Tuk-Tuk Tracker API
  version: 1.0.0
  description: |
    ## Real-Time Three-Wheeler (Tuk-Tuk) Tracking & Movement Logging System
    Developed for **Sri Lanka Police** to support operational visibility and investigations.

  contact:
    name: Sri Lanka Police IT Division
  license:
    name: MIT
servers:
  - url: /api/v1
    description: API Version 1

#security scheme
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Enter your JWT token obtained from POST /auth/login.
        Format: `Bearer <token>`

#these are reusable schemas
  schemas:
...
    UpdateProvinceRequest:
      type: object
      properties:
        name: { type: string }
        code: { type: string }

    #for district section
...
    UpdateDistrictRequest:
      type: object
      properties:
        name: { type: string }
        province: { type: string }
        code: { type: string }

    #for station section
...
    CreateStationRequest:
      type: object
      required:
        - name
        - stationCode
        - district
        - province
      properties:
        name:
          type: string
          example: Negombo Police Station
        stationCode:
          type: string
          example: NEG-102
        district:
          type: string
          example: Gampaha
        province:
          type: string
          example: Western
        address:
          type: string
          example: Main Street, Negombo
        contactNumber:
          type: string
          example: 031-2222222
    UpdateStationRequest:
      type: object
      properties:
        name: { type: string }
        stationCode: { type: string }
        district: { type: string }
        province: { type: string }
        address: { type: string }
        contactNumber: { type: string }

    #for driver section
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Description of what went wrong
...
  responses:
    Unauthorized:
      description: No token provided or token is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Access denied. Please log in to continue.
    Forbidden:
      description: Token valid but role does not have permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Your role 'station_officer' does not have permission to perform this action.
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Resource not found.
...
security:
  - bearerAuth: []

#tags for each section
...
paths:
...
  /master/provinces:
    get:
      summary: Get all 9 provinces
      operationId: getProvinces
      tags:
        - Master Data Management
      responses:
        '200':
          description: All provinces
          content:
            application/json:
              example:
                success: true
                results: 9
                data:
                  - name: Western
                    code: WP
                  - name: Central
                    code: CP
        '401':
          $ref: '#/components/responses/Unauthorized'
  /master/provinces/{name}:
    get:
      summary: Get a province by name
      operationId: getProvince
      tags:
        - Master Data Management
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
          example: Western
      responses:
        '200':
          description: Province found
        '404':
          description: Province not found
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      summary: Update a province (hq_admin and province officer only)
      operationId: updateProvince
      tags:
        - Master Data Management
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
          example: Western
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProvinceRequest'
            example:
              name: Western
              code: WP
      responses:
        '200':
          description: Province updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      summary: Delete a province (hq_admin only)
      description: |
        Only hq_admin can delete a province.
      operationId: deleteProvince
      tags:
        - Master Data Management
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
          example: Western
      responses:
        '200':
          description: Province deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /master/districts:
    get:
      summary: Get all 25 districts (hq_admin, province officer, station_officer only)
      description: Filter by province name to get districts in one province.
      operationId: getDistricts
      tags:
        - Master Data Management
      parameters:
        - in: query
          name: province
          schema:
            type: string
          example: Western
      responses:
        '200':
          description: List of districts
          content:
            application/json:
              example:
                success: true
                results: 3
                data:
                  - name: Colombo
                    province:
                      name: Western
        '401':
          $ref: '#/components/responses/Unauthorized'
  /master/districts/{name}:
    get:
      summary: Get a district by name (hq_admin, province officer, station_officer only)
      operationId: getDistrict
      tags:
        - Master Data Management
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
          example: Colombo
      responses:
        '200':
          description: District found
        '404':
          description: District not found
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    patch:
      summary: Update a district (hq_admin only)
      operationId: updateDistrict
      tags:
        - Master Data Management
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
          example: Colombo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDistrictRequest'
            example:
              name: Colombo
              province: Western
              code: COL
      responses:
        '200':
          description: District updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a district (hq_admin only)
      description: |
        Only hq_admin can delete a district.
      operationId: deleteDistrict
      tags:
        - Master Data Management
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
          example: Colombo
      responses:
        '200':
          description: District deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /master/stations:
    post:
      summary: Create a new police station (hq_admin and station_officer only)
      operationId: createStation
      tags:
        - Master Data Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStationRequest'
      responses:
        '201':
          description: Station created
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      summary: Get all police stations
      description: Filter by province name or district name.
      operationId: getStations
      tags:
        - Master Data Management
      parameters:
        - in: query
          name: district
          schema:
            type: string
          example: Colombo
        - in: query
          name: province
          schema:
            type: string
          example: Western
      responses:
        '200':
          description: List of police stations
          content:
            application/json:
              example:
                success: true
                results: 3
                data:
                  - name: Colombo Central Police Station
                    stationCode: COL-101
        '401':
          $ref: '#/components/responses/Unauthorized'
  /master/stations/{code}:
    get:
      summary: Get a police station by code
      operationId: getStation
      tags:
        - Master Data Management
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
          example: COL-101
      responses:
        '200':
          description: Police station found
        '404':
          description: Police station not found
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    patch:
      summary: Update a police station (hq_admin and station_officer only)
      operationId: updateStation
      tags:
        - Master Data Management
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
          example: COL-101
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStationRequest'
      responses:
        '200':
          description: Station updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      summary: Delete a police station (hq_admin only)
      operationId: deleteStation
      tags:
        - Master Data Management
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
          example: COL-101
      responses:
        '204':
          description: Station deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'